+++ src/main/java/org/codeforamerica/shiba/mnit/FilenetWebServiceClient.java
@@ -15,21 +15,14 @@ import java.time.format.DateTimeFormatter;
 import java.util.List;
 import javax.activation.DataHandler;
 import javax.mail.util.ByteArrayDataSource;
+import javax.xml.bind.JAXBElement;
 import javax.xml.namespace.QName;
-import javax.xml.soap.SOAPElement;
-import javax.xml.soap.SOAPException;
-import javax.xml.soap.SOAPHeader;
-import javax.xml.soap.SOAPHeaderElement;
-import javax.xml.soap.SOAPMessage;
+import javax.xml.soap.*;
 import lombok.extern.slf4j.Slf4j;
 import org.codeforamerica.shiba.CountyMap;
 import org.codeforamerica.shiba.application.ApplicationRepository;
 import org.codeforamerica.shiba.application.FlowType;
-import org.codeforamerica.shiba.esbwsdl.CmisContentStreamType;
-import org.codeforamerica.shiba.esbwsdl.CmisPropertiesType;
-import org.codeforamerica.shiba.esbwsdl.CmisProperty;
-import org.codeforamerica.shiba.esbwsdl.CmisPropertyString;
-import org.codeforamerica.shiba.esbwsdl.CreateDocument;
+import org.codeforamerica.shiba.filenetwsdl.*;
 import org.codeforamerica.shiba.output.ApplicationFile;
 import org.codeforamerica.shiba.output.Document;
 import org.jetbrains.annotations.NotNull;
@@ -46,23 +39,25 @@ import org.springframework.ws.soap.saaj.SaajSoapMessage;
 
 @Component
 @Slf4j
-public class MnitEsbWebServiceClient {
+public class FilenetWebServiceClient {
 
-  private final WebServiceTemplate webServiceTemplate;
+  private static final String APPLICATION_PDF = "application/pdf";
+  private static final String APPLICATION_XML = "application/xml";
+  private final WebServiceTemplate filenetWebServiceTemplate;
   private final Clock clock;
   private final String username;
   private final String password;
   private final ApplicationRepository applicationRepository;
 
   @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
-  public MnitEsbWebServiceClient(
-      @Qualifier("webServiceTemplate") WebServiceTemplate webServiceTemplate,
+  public FilenetWebServiceClient(
+      @Qualifier("filenetWebServiceTemplate") WebServiceTemplate webServiceTemplate,
       Clock clock,
-      @Value("${mnit-esb.username}") String username,
-      @Value("${mnit-esb.password}") String password,
+      @Value("${mnit-filenet.username}") String username,
+      @Value("${mnit-filenet.password}") String password,
       CountyMap<CountyRoutingDestination> countyMap,
       ApplicationRepository applicationRepository) {
-    this.webServiceTemplate = webServiceTemplate;
+    this.filenetWebServiceTemplate = webServiceTemplate;
     this.clock = clock;
     this.username = username;
     this.password = password;
@@ -71,28 +66,26 @@ public class MnitEsbWebServiceClient {
 
   @Retryable(
       value = {Exception.class},
-      maxAttemptsExpression = "#{${mnit-esb.max-attempts}}",
+      maxAttemptsExpression = "#{${mnit-filenet.max-attempts}}",
       backoff = @Backoff(
-          delayExpression = "#{${mnit-esb.delay}}",
-          multiplierExpression = "#{${mnit-esb.multiplier}}",
-          maxDelayExpression = "#{${mnit-esb.max-delay}}"
+          delayExpression = "#{${mnit-filenet.delay}}",
+          multiplierExpression = "#{${mnit-filenet.multiplier}}",
+          maxDelayExpression = "#{${mnit-filenet.max-delay}}"
       ),
       listeners = {"esbRetryListener"}
   )
   public void send(ApplicationFile applicationFile, RoutingDestination routingDestination,
-      String applicationNumber, Document applicationDocument, FlowType flowType) {
-    MDC.put("applicationId", applicationNumber);
+      String applicationNumber,
+      Document applicationDocument, FlowType flowType) {
     MDC.put("applicationFile", applicationFile.getFileName());
     CreateDocument createDocument = new CreateDocument();
-    createDocument.setFolderId("workspace://SpacesStore/" + routingDestination.getFolderId());
-    createDocument.setRepositoryId("<Unknown");
-    createDocument.setTypeId("document");
+    createDocument.setRepositoryId("Programs");
     setPropertiesOnDocument(applicationFile, routingDestination, applicationNumber,
         applicationDocument,
         flowType, createDocument);
     setContentStreamOnDocument(applicationFile, createDocument);
 
-    webServiceTemplate.marshalSendAndReceive(createDocument, message -> {
+    filenetWebServiceTemplate.marshalSendAndReceive(createDocument, message -> {
       SOAPMessage soapMessage = ((SaajSoapMessage) message).getSaajMessage();
       try {
         SOAPHeader soapHeader = soapMessage.getSOAPHeader();
@@ -125,10 +118,10 @@ public class MnitEsbWebServiceClient {
             flowType);
       }
     });
+
     applicationRepository.updateStatus(applicationNumber, applicationDocument, DELIVERED);
   }
 
-  // Recover method has to have the same arguments at the retryable
   @Recover
   public void logErrorToSentry(Exception e, ApplicationFile applicationFile,
       RoutingDestination routingDestination,
@@ -137,12 +130,84 @@ public class MnitEsbWebServiceClient {
     log.error("Application failed to send: " + applicationFile.getFileName(), e);
   }
 
+
+  private void setPropertiesOnDocument(ApplicationFile applicationFile,
+      RoutingDestination routingDestination, String applicationNumber,
+      Document applicationDocument, FlowType flowType,
+      CreateDocument createDocument) {
+    CmisPropertiesType properties = new CmisPropertiesType();
+    List<CmisProperty> propertiesList = properties.getProperty();
+
+    CmisPropertyBoolean read = createCmisPropertyBoolean("Read", false);
+    CmisPropertyString originalFileName = createCmisPropertyString("OriginalFileName",
+        applicationFile.getFileName());
+    CmisPropertyString cmisName = createCmisPropertyString("cmis:name",
+        applicationFile.getFileName());
+    CmisPropertyString fileType = createCmisPropertyString("FileType", "Misc");
+    CmisPropertyString npi = createCmisPropertyString("NPI", routingDestination.getDhsProviderId());
+    CmisPropertyString mnitsMailboxTransactionType = createCmisPropertyString(
+        "MNITSMailboxTransactionType", "OLA");
+    CmisPropertyString description = createCmisPropertyString("Description",
+        generateDocumentDescription(applicationFile, applicationDocument, applicationNumber,
+            flowType));
+    CmisPropertyString source = createCmisPropertyString("Source", "MNITS");
+    CmisPropertyString flow = createCmisPropertyString("Flow", "Inbound");
+    CmisPropertyId cmisObjectTypeId = createCmisPropertyId("cmis:objectTypeId", "MNITSMailbox");
+
+    propertiesList
+        .addAll(
+            List.of(read, originalFileName, cmisName, fileType, npi, mnitsMailboxTransactionType,
+                source, flow, cmisObjectTypeId, description));
+    createDocument.setProperties(properties);
+  }
+
   @NotNull
-  private CmisPropertyString createCmisPropertyString(String property, String value) {
-    CmisPropertyString fileNameProperty = new CmisPropertyString();
-    fileNameProperty.setName(property);
-    fileNameProperty.setValue(value);
-    return fileNameProperty;
+  private CmisPropertyString createCmisPropertyString(String propertyDefinitionId,
+      String propertyValue) {
+    CmisPropertyString stringProperty = new CmisPropertyString();
+    stringProperty.setPropertyDefinitionId(propertyDefinitionId);
+    stringProperty.getValue().add(propertyValue);
+    return stringProperty;
+  }
+
+  @NotNull
+  private CmisPropertyBoolean createCmisPropertyBoolean(String propertyDefinitionId,
+      Boolean propertyValue) {
+    CmisPropertyBoolean booleanProperty = new CmisPropertyBoolean();
+    booleanProperty.setPropertyDefinitionId(propertyDefinitionId);
+    booleanProperty.getValue().add(propertyValue);
+    return booleanProperty;
+  }
+
+  @NotNull
+  private CmisPropertyId createCmisPropertyId(String propertyDefinitionId, String propertyValue) {
+    CmisPropertyId idProperty = new CmisPropertyId();
+    idProperty.setPropertyDefinitionId(propertyDefinitionId);
+    idProperty.getValue().add(propertyValue);
+    return idProperty;
+  }
+
+  private void setContentStreamOnDocument(ApplicationFile applicationFile,
+      CreateDocument createDocument) {
+
+    String fileName = applicationFile.getFileName();
+    String mimeType = MimeTypeUtils.APPLICATION_OCTET_STREAM_VALUE;
+    if (fileName.toLowerCase().endsWith(".pdf")) {
+      mimeType = APPLICATION_PDF;
+    } else if (fileName.toLowerCase().endsWith(".xml")) {
+      mimeType = APPLICATION_XML;
+    }
+
+    CmisContentStreamType contentStream = new CmisContentStreamType();
+    contentStream.setLength(BigInteger.ZERO);
+    contentStream.setStream(new DataHandler(new ByteArrayDataSource(applicationFile.getFileBytes(),
+        mimeType)));
+
+    ObjectFactory ob = new ObjectFactory();
+    JAXBElement<CmisContentStreamType> jaxbContentStream = ob.createCreateDocumentContentStream(
+        contentStream);
+    jaxbContentStream.getValue().setMimeType(mimeType);
+    createDocument.setContentStream(jaxbContentStream);
   }
 
   @NotNull
@@ -168,33 +233,5 @@ public class MnitEsbWebServiceClient {
       }
     }
     return docDescription;
-  }
-
-  private void setPropertiesOnDocument(ApplicationFile applicationFile,
-      RoutingDestination routingDestination,
-      String applicationNumber, Document applicationDocument, FlowType flowType,
-      CreateDocument createDocument) {
-    CmisPropertiesType properties = new CmisPropertiesType();
-    List<CmisProperty> propertyUris = properties.getPropertyUriOrPropertyIdOrPropertyString();
-    CmisPropertyString fileNameProperty = createCmisPropertyString("Name",
-        applicationFile.getFileName());
-    CmisPropertyString subject = createCmisPropertyString("subject", "MN Benefits Application");
-    CmisPropertyString description = createCmisPropertyString("description",
-        generateDocumentDescription(applicationFile, applicationDocument, applicationNumber,
-            flowType));
-    CmisPropertyString dhsProviderId = createCmisPropertyString("dhsProviderId",
-        routingDestination.getDhsProviderId());
-    propertyUris
-        .addAll(List.of(fileNameProperty, subject, description, description, dhsProviderId));
-    createDocument.setProperties(properties);
-  }
-
-  private void setContentStreamOnDocument(ApplicationFile applicationFile,
-      CreateDocument createDocument) {
-    CmisContentStreamType contentStream = new CmisContentStreamType();
-    contentStream.setLength(BigInteger.ZERO);
-    contentStream.setStream(new DataHandler(new ByteArrayDataSource(applicationFile.getFileBytes(),
-        MimeTypeUtils.APPLICATION_OCTET_STREAM_VALUE)));
-    createDocument.setContentStream(contentStream);
   }
 }
